Remove unused code and files
