+++ app/src/main/java/dev/arkbuilders/arkmemo/ui/viewmodels/QRViewModel.kt
@@ -1,6 +1,5 @@
 package dev.arkbuilders.arkmemo.ui.viewmodels
 
-import android.annotation.SuppressLint
 import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.Color
@@ -12,7 +11,7 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dev.arkbuilders.arkmemo.di.IO_DISPATCHER
-import dev.arkbuilders.arkmemo.utils.dp2Px
+import dev.arkbuilders.arkmemo.utils.dpToPx
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -22,17 +21,16 @@ import javax.inject.Inject
 import javax.inject.Named
 
 
-@SuppressLint("StaticFieldLeak")
 @HiltViewModel
 class QRViewModel @Inject constructor(
     @Named(IO_DISPATCHER) private val iODispatcher: CoroutineDispatcher,
-    @ApplicationContext private val context: Context,
+    @ApplicationContext private val appContext: Context,
 ) : ViewModel() {
 
     fun generateQRCode(text: String, onSuccess: (bitmap: Bitmap) -> Unit) {
         viewModelScope.launch(iODispatcher) {
             // Initializing the QR Encoder with your value to be encoded, type you required and Dimension
-            val qrgEncoder = QRGEncoder(text, null, QRGContents.Type.TEXT, 300.dp2Px())
+            val qrgEncoder = QRGEncoder(text, null, QRGContents.Type.TEXT, 300.dpToPx())
             qrgEncoder.colorBlack = Color.BLACK
             qrgEncoder.colorWhite = Color.WHITE
             withContext(Dispatchers.Main) {
@@ -46,7 +44,7 @@ class QRViewModel @Inject constructor(
             // Save with location, value, bitmap returned and type of Image(JPG/PNG).
             val qrgSaver = QRGSaver()
 
-            val savePath = (context.getExternalFilesDir(null)?.path + "/images/").apply {
+            val savePath = (appContext.getExternalFilesDir(null)?.path + "/images/").apply {
                 File(this).mkdirs()
             }
