+++ app/src/main/java/dev/arkbuilders/arkmemo/utils/DimensionUtils.kt
@@ -3,7 +3,7 @@ package dev.arkbuilders.arkmemo.utils
 import android.content.res.Resources
 import android.util.TypedValue
 
-fun Int.dp2Px(): Int {
+fun Int.dpToPx(): Int {
     return TypedValue.applyDimension(
         TypedValue.COMPLEX_UNIT_DIP,
         this.toFloat(),
@@ -11,7 +11,7 @@ fun Int.dp2Px(): Int {
     ).toInt()
 }
 
-fun Float.dp2Px(): Float {
+fun Float.dpToPx(): Float {
     return TypedValue.applyDimension(
         TypedValue.COMPLEX_UNIT_DIP,
         this,
