+++ app/src/main/res/drawable/bg_round_button.xml
@@ -1,8 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="oval">
-
-<!--    <corners android:radius="@dimen/common_button_corner_radius"/>-->
     <solid android:color="@color/warning"/>
 
 </shape>
