+++ app/src/main/res/layout/fragment_home.xml
@@ -107,21 +107,6 @@
         android:background="@drawable/bg_rect_button"
         android:contentDescription="@string/ark_memo_record" />
 
-<!--    <ImageView-->
-<!--        android:layout_width="@dimen/common_button_height"-->
-<!--        android:layout_height="@dimen/common_button_height"-->
-<!--        app:layout_constraintStart_toEndOf="@+id/iv_record"-->
-<!--        app:layout_constraintEnd_toStartOf="@+id/iv_draw"-->
-<!--        android:layout_marginStart="8dp"-->
-<!--        app:layout_constraintTop_toTopOf="@+id/iv_record"-->
-<!--        app:layout_constraintBottom_toBottomOf="@+id/iv_record"-->
-<!--        android:src="@drawable/ic_attach"-->
-<!--        android:padding="14dp"-->
-<!--        android:id="@+id/iv_attach_file"-->
-<!--        android:visibility="gone"-->
-<!--        app:layout_constraintHorizontal_chainStyle="packed"-->
-<!--        android:background="@drawable/bg_rect_button"/>-->
-
     <ImageView
         android:layout_width="@dimen/common_button_height"
         android:layout_height="@dimen/common_button_height"
@@ -136,21 +121,6 @@
         android:id="@+id/iv_draw"
         android:background="@drawable/bg_rect_button"
         android:contentDescription="@string/hint_new_graphical_note" />
-
-<!--    <ImageView-->
-<!--        android:layout_width="@dimen/common_button_height"-->
-<!--        android:layout_height="@dimen/common_button_height"-->
-<!--        app:layout_constraintStart_toEndOf="@+id/iv_draw"-->
-<!--        android:layout_marginStart="8dp"-->
-<!--        app:layout_constraintTop_toTopOf="@+id/iv_record"-->
-<!--        app:layout_constraintBottom_toBottomOf="@+id/iv_record"-->
-<!--        app:layout_constraintEnd_toEndOf="parent"-->
-<!--        app:layout_constraintHorizontal_chainStyle="packed"-->
-<!--        android:src="@drawable/ic_photo"-->
-<!--        android:padding="14dp"-->
-<!--        android:visibility="gone"-->
-<!--        android:id="@+id/iv_attach_media"-->
-<!--        android:background="@drawable/bg_rect_button"/>-->
 
     <TextView
         android:text="@string/ark_memo_paste"
@@ -271,19 +241,6 @@
 
     </androidx.core.widget.NestedScrollView>
 
-<!--    <com.google.android.material.floatingactionbutton.FloatingActionButton-->
-<!--        android:id="@+id/fab_attach_media"-->
-<!--        android:layout_width="wrap_content"-->
-<!--        android:layout_height="wrap_content"-->
-<!--        app:fabCustomSize="@dimen/fab_size"-->
-<!--        android:src="@drawable/ic_photo"-->
-<!--        app:layout_constraintEnd_toEndOf="@+id/fab_new_action"-->
-<!--        app:layout_constraintStart_toStartOf="@+id/fab_new_action"-->
-<!--        app:layout_constraintBottom_toTopOf="@+id/fab_new_action"-->
-<!--        android:visibility="gone"-->
-<!--        tools:visibility="visible"-->
-<!--        android:layout_marginBottom="@dimen/common_padding" />-->
-
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/fab_draw"
         android:layout_width="wrap_content"
@@ -297,19 +254,6 @@
         tools:visibility="visible"
         android:layout_marginBottom="@dimen/common_padding"
         android:contentDescription="@string/hint_new_graphical_note" />
-
-<!--    <com.google.android.material.floatingactionbutton.FloatingActionButton-->
-<!--        android:id="@+id/fab_attach_file"-->
-<!--        android:layout_width="wrap_content"-->
-<!--        android:layout_height="wrap_content"-->
-<!--        app:fabCustomSize="@dimen/fab_size"-->
-<!--        android:src="@drawable/ic_attach"-->
-<!--        app:layout_constraintEnd_toEndOf="@+id/fab_new_action"-->
-<!--        app:layout_constraintStart_toStartOf="@+id/fab_new_action"-->
-<!--        app:layout_constraintBottom_toTopOf="@+id/fab_draw"-->
-<!--        android:visibility="gone"-->
-<!--        tools:visibility="visible"-->
-<!--        android:layout_marginBottom="@dimen/common_padding" />-->
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/fab_record_voice"
@@ -360,17 +304,6 @@
         android:clickable="true"
         app:layout_constraintBottom_toTopOf="@+id/divider_bottom_control"/>
 
-<!--    <TextView-->
-<!--        android:layout_width="wrap_content"-->
-<!--        android:layout_height="wrap_content"-->
-<!--        style="@style/TextLarge"-->
-<!--        app:layout_constraintEnd_toStartOf="@+id/fab_attach_media"-->
-<!--        app:layout_constraintTop_toTopOf="@+id/fab_attach_media"-->
-<!--        app:layout_constraintBottom_toBottomOf="@+id/fab_attach_media"-->
-<!--        android:text="@string/photo_and_video"-->
-<!--        android:layout_marginEnd="@dimen/fab_text_margin"-->
-<!--        android:id="@+id/tv_attach_media"/>-->
-
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -381,17 +314,6 @@
         android:text="@string/fab_create_drawing_text"
         android:layout_marginEnd="@dimen/fab_text_margin"
         android:id="@+id/tv_create_drawing"/>
-
-<!--    <TextView-->
-<!--        android:layout_width="wrap_content"-->
-<!--        android:layout_height="wrap_content"-->
-<!--        style="@style/TextLarge"-->
-<!--        app:layout_constraintEnd_toStartOf="@+id/fab_attach_file"-->
-<!--        app:layout_constraintTop_toTopOf="@+id/fab_attach_file"-->
-<!--        app:layout_constraintBottom_toBottomOf="@+id/fab_attach_file"-->
-<!--        android:text="@string/fab_attach_file"-->
-<!--        android:layout_marginEnd="@dimen/fab_text_margin"-->
-<!--        android:id="@+id/tv_attach_file"/>-->
 
     <TextView
         android:layout_width="wrap_content"
