--- app/src/main/res/menu/main_menu.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-<item
-    android:id="@+id/settings"
-    android:title="@string/settings"
-    android:icon="@drawable/settings"
-    app:showAsAction="ifRoom"
-    />
-</menu>
