+++ app/src/main/res/values/strings.xml
@@ -12,7 +12,6 @@
     <string name="new_note">New note</string>
     <string name="note">Note</string>
     <string name="date">Date</string>
-    <string name="settings">Settings</string>
     <string name="about">About</string>
     <string name="path_pref_key">Path</string>
     <string name="choose_folder">Choose folder to save your notes</string>
