--- app/src/main/res/xml/root_preferences.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen xmlns:app="http://schemas.android.com/apk/res-auto">
-   <PreferenceCategory app:title="@string/your_notes">
-       <dev.arkbuilders.arkmemo.ui.views.PathPreference
-           app:key="@string/path_pref_key"
-           app:layout="@layout/path_pref" />
-   </PreferenceCategory>
-</PreferenceScreen>
